refactor: rename contains resolvers and schemas file
