+++ src/FileMerger.ts
